Add files via upload
This is an exploratory analysis of  Census data from the United States 
in the United States. This dataset contains population data for countries and states in the
US from 2010 to 2015.
From the analysis i was able to answer the following questions
1. Which state has the most countries in it? (hint: consider the sumlevel key carefully! You'll need this for
future questions too...)
2. Only looking at the three most populous countries for each state, what are the three most populous 
states (in order of highest population to lowest population)? Use CENSUS2010POP
3. Which city has the most countries in it?
4. Which region has the most division in it?